+++ coretech/messaging/shared/UdpServer.h
@@ -23,8 +23,8 @@ public:
   
   int GetNumClients();
 
-  ssize_t Send(const char* data, int size);
-  ssize_t Recv(char* data, int maxSize);
+  ssize_t Send(const char* data, size_t size);
+  ssize_t Recv(char* data, size_t maxSize);
   
 private:
