+++ engine/cozmoAPI/comms/gameComms.h
@@ -82,7 +82,7 @@ namespace Cozmo {
     const char*    advertisementRegIP_;
     int            advertisementRegPort_;
     
-    static const int MAX_RECV_BUF_SIZE = 1920000; // [TODO] 1.9MB seems excessive?
+    static const size_t MAX_RECV_BUF_SIZE = 1920000; // [TODO] 1.9MB seems excessive?
     u8  _recvBuf[MAX_RECV_BUF_SIZE];
     ssize_t recvDataSize = 0;
